ahora se pueden añadir comics

+++ app/Database/Migrations/2023-10-24-070048_TransactionInfo.php
@@ -23,7 +23,7 @@ class TransactionInfo extends Migration
                 "constraint"=>"255",
             ],
             "price"=> [
-                "type"=>"INT",
+                "type"=>"FLOAT",
             ],
             'created_at timestamp default current_timestamp',
             'updated_at timestamp default current_timestamp on update current_timestamp'
